% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{add_geography_names}
\alias{add_geography_names}
\title{Join Geography Names}
\usage{
add_geography_names(data, dataset_id = "NM_1_1")
}
\arguments{
\item{data}{Data frame with GEOGRAPHY_CODE column}

\item{dataset_id}{Dataset to get geography names from}
}
\value{
Data frame with GEOGRAPHY_NAME added
}
\description{
Adds human-readable geography names to data.
}
\examples{
\donttest{
data <- fetch_nomis("NM_1_1", time = "latest", geography = "TYPE499")
data_with_names <- add_geography_names(data)
}
}
