% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovery.R
\name{dataset_overview}
\alias{dataset_overview}
\title{Get Dataset Overview}
\usage{
dataset_overview(id, select = NULL)
}
\arguments{
\item{id}{Dataset ID (required)}

\item{select}{Character vector of sections to return}
}
\value{
Tibble with overview information
}
\description{
Get Dataset Overview
}
\examples{
\donttest{
dataset_overview("NM_1_1")
dataset_overview("NM_1_1", select = c("Keywords", "Units"))
}
}
