% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive.R
\name{explore_dataset}
\alias{explore_dataset}
\title{Interactive Dataset Explorer}
\usage{
explore_dataset(id)
}
\arguments{
\item{id}{Dataset ID}
}
\value{
Selected codes as a list
}
\description{
Opens an interactive menu to explore dataset dimensions and codes.
}
\examples{
\dontrun{
explore_dataset("NM_1_1")
}
}
