% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{fetch_codelist}
\alias{fetch_codelist}
\title{Fetch Codelist}
\usage{
fetch_codelist(id, concept, search = NULL)
}
\arguments{
\item{id}{Dataset ID}

\item{concept}{Concept name}

\item{search}{Search term}
}
\value{
Tibble of codes
}
\description{
Fetch Codelist
}
\examples{
\donttest{
fetch_codelist("NM_1_1", "geography")
fetch_codelist("NM_1_1", "geography", "*manchester*")
}
}
