% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomisdata-package.R
\docType{package}
\name{nomisdata-package}
\alias{nomisdata}
\alias{nomisdata-package}
\title{nomisdata: Access UK Labour Market Data}
\description{
Provides tools to discover, query, and download UK official statistics
from the Nomis API. Includes Census data, Labour Force Survey, benefit
statistics, and economic/demographic data.
}
\details{
\strong{Note}: This package is not affiliated with the Office for National
Statistics or the University of Durham. It is an independent implementation
of the Nomis API.

Main functions:
\itemize{
\item \code{fetch_nomis()}: Download data
\item \code{search_datasets()}: Find datasets
\item \code{describe_dataset()}: Get dataset structure
\item \code{get_codes()}: Get dimension codes
\item \code{lookup_geography()}: Find geography codes
}

For more information: https://www.nomisweb.co.uk/api/v01/help
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cherylisabella/nomisdata}
  \item Report bugs at \url{https://github.com/cherylisabella/nomisdata/issues}
}

}
\author{
\strong{Maintainer}: Cheryl Isabella Lim \email{cheryl.academic@gmail.com} (\href{https://orcid.org/0009-0004-5766-1392}{ORCID})

}
\keyword{internal}
