% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovery.R
\name{search_datasets}
\alias{search_datasets}
\title{Search for Datasets}
\usage{
search_datasets(
  name = NULL,
  keywords = NULL,
  description = NULL,
  content_type = NULL,
  units = NULL
)
}
\arguments{
\item{name}{Character vector of name search terms (supports wildcards *)}

\item{keywords}{Character vector of keyword search terms}

\item{description}{Character vector of description search terms}

\item{content_type}{Character vector of content types}

\item{units}{Character vector of units}
}
\value{
Tibble of matching datasets
}
\description{
Search for Datasets
}
\examples{
\donttest{
search_datasets(name = "*employment*")
search_datasets(keywords = "census")
search_datasets(name = "*benefit*", keywords = "claimants")
}
}
