% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{tidy_names}
\alias{tidy_names}
\title{Tidy Column Names}
\usage{
tidy_names(df, style = "snake_case")
}
\arguments{
\item{df}{Data frame}

\item{style}{Naming style: "snake_case", "camelCase", "period.case"}
}
\value{
Data frame with tidied names
}
\description{
Tidy Column Names
}
\examples{
df <- data.frame(GEOGRAPHY_NAME = "UK", OBS_VALUE = 100)
tidy_names(df)
}
