% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{enable_cache}
\alias{enable_cache}
\title{Enable disk caching}
\usage{
enable_cache(path = NULL)
}
\arguments{
\item{path}{Directory for cache. Defaults to user cache directory.}
}
\value{
Invisible TRUE
}
\description{
Enable disk caching
}
\examples{
\dontrun{
enable_cache()
# Custom location
enable_cache("~/my_nomis_cache")
}
}
